THUNDER ATTACK : work not properly done but thunder. -> proper work
