+++ API_SoftDesk/projects/urls.py
@@ -5,14 +5,18 @@ from django.urls import path
 from rest_framework.routers import (SimpleRouter,)
 # Local Libs
 from . import views
+from .crud_project import ProjectCRUD
+from .crud_user import UserCRUD
+from .crud_issue import IssueCRUD
+from .crud_comment import CommentCRUD
 
 #app_name = ""
 
 router = SimpleRouter()
-router.register(r"", views.ProjectCRUD, basename="projects")
-router.register(r"^(?P<id>[^/.]+)/users", views.UserCRUD, basename="users")
-router.register(r"^(?P<id>[^/.]+)/issues", views.IssueCRUD, basename="issues")
-router.register(r"^(?P<id>[^/.]+)/issues/(?P<issue_id>[^/.]+)/comments", views.CommentCRUD, basename="comments")
+router.register(r"", ProjectCRUD, basename="projects")
+router.register(r"^(?P<id>[^/.]+)/users", UserCRUD, basename="users")
+router.register(r"^(?P<id>[^/.]+)/issues", IssueCRUD, basename="issues")
+router.register(r"^(?P<id>[^/.]+)/issues/(?P<issue_id>[^/.]+)/comments", CommentCRUD, basename="comments")
 print("------------------URLS--------------")
 for url in router.urls:
     print(f"url : {url}")
